move launcher related function into job_launcher
