+++ src/ClusterManager/test_job_deployer.py
@@ -6,7 +6,7 @@ import random
 import time
 from kubernetes.client.rest import ApiException
 
-from job_deployer import JobDeployer
+from job_launcher import JobDeployer
 
 import logging
 logging.basicConfig(
@@ -43,7 +43,7 @@ spec:
         body = yaml.full_load(raw_yaml)
 
         # with self.assertRaises(ApiException):
-        job_deployer.create_pod(body)
+        job_deployer._create_pod(body)
 
     def test_delete_pod(self):
         pod_name = ''.join(random.choice(string.ascii_lowercase + string.digits) for _ in range(16))
@@ -51,13 +51,13 @@ spec:
 
         job_deployer = self.create_job_deployer()
 
-        job_deployer.delete_pod(pod_name)
+        job_deployer._delete_pod(pod_name)
 
     def test_cleanup_pods(self):
         job_deployer = self.create_job_deployer()
         pod_names = ["pod-1", "pod-2"]
 
-        job_deployer.cleanup_pods(pod_names)
+        job_deployer._cleanup_pods(pod_names)
 
     def test_get_pod_by_label(self):
         job_deployer = self.create_job_deployer()
@@ -71,7 +71,7 @@ spec:
         job_deployer = self.create_job_deployer()
         label_selector = "run=some_job_id"
 
-        services = job_deployer.get_services_by_label(label_selector)
+        services = job_deployer._get_services_by_label(label_selector)
 
         self.assertEqual(0, len(services))
 
@@ -132,4 +132,4 @@ spec:
         status_code, ouput = job_deployer.pod_exec(pod_name, bad_command, 1)
         self.assertEqual(-1, status_code)
 
-        job_deployer.delete_pod(pod_name)
+        job_deployer._delete_pod(pod_name)
